Chat is chat, Command is command.
