+++ src/main/java/cn/nukkit/Player.java
@@ -2475,22 +2475,20 @@ public class Player extends Human implements CommandSender, InventoryHolder, Chu
                     textPacket.message = this.removeFormat ? TextFormat.clean(textPacket.message) : textPacket.message;
                     for (String msg : textPacket.message.split("\n")) {
                         if (!"".equals(msg.trim()) && msg.length() <= 255 && this.messageCounter-- > 0) {
-                            PlayerCommandPreprocessEvent commandPreprocessEvent = new PlayerCommandPreprocessEvent(this, msg);
-
-                            if (commandPreprocessEvent.getMessage().length() > 320) {
-                                ev.setCancelled();
-                            }
-                            this.server.getPluginManager().callEvent(commandPreprocessEvent);
-
-                            if (commandPreprocessEvent.isCancelled()) {
-                                break;
-                            }
-                            if (commandPreprocessEvent.getMessage().startsWith("/")) { //Command
+                            if (msg.startsWith("/")) { //Command
+                                PlayerCommandPreprocessEvent commandPreprocessEvent = new PlayerCommandPreprocessEvent(this, msg);
+                                if (commandPreprocessEvent.getMessage().length() > 320) {
+                                    ev.setCancelled();
+                                }
+                                this.server.getPluginManager().callEvent(commandPreprocessEvent);
+                                if (commandPreprocessEvent.isCancelled()) {
+                                    break;
+                                }
                                 //Timings::playerCommandTimer.startTiming();
                                 this.server.dispatchCommand(commandPreprocessEvent.getPlayer(), commandPreprocessEvent.getMessage().substring(1));
                                 //Timings::playerCommandTimer.stopTiming();
-                            } else {
-                                PlayerChatEvent chatEvent = new PlayerChatEvent(this, commandPreprocessEvent.getMessage());
+                            } else { //Chat
+                                PlayerChatEvent chatEvent = new PlayerChatEvent(this, msg);
                                 this.server.getPluginManager().callEvent(chatEvent);
                                 if (!chatEvent.isCancelled()) {
                                     this.server.broadcastMessage(this.getServer().getLanguage().translateString(chatEvent.getFormat(), new String[]{chatEvent.getPlayer().getDisplayName(), chatEvent.getMessage()}), chatEvent.getRecipients());
