+++ src/main/java/cn/nukkit/event/player/PlayerCommandPreprocessEvent.java
@@ -4,26 +4,15 @@ import cn.nukkit.Player;
 import cn.nukkit.event.Cancellable;
 import cn.nukkit.event.HandlerList;
 
-public class PlayerCommandPreprocessEvent extends PlayerEvent implements Cancellable {
+public class PlayerCommandPreprocessEvent extends PlayerMessageEvent implements Cancellable {
     private static final HandlerList handlers = new HandlerList();
 
     public static HandlerList getHandlers() {
         return handlers;
     }
 
-    protected String message;
-
-
     public PlayerCommandPreprocessEvent(Player player, String message) {
         this.player = player;
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return this.message;
-    }
-
-    public void setMessage(String message) {
         this.message = message;
     }
